minor package refactoring and locking down the mock
