+++ sync/src/main/resources/application.properties
@@ -26,6 +26,7 @@ secret.enabled=true
 # test overrides
 %test.sso.enabled=false
 %test.secret.enabled=false
+%test.sync.run-control-plane-simulation=true
 
 # control plane properties
 control-plane/mp-rest/url=${control-plane.url}
