+++ sync/src/test/java/org/bf2/sync/DirectLocalLookup.java
@@ -7,7 +7,6 @@ import javax.inject.Inject;
 
 import org.bf2.operator.resources.v1alpha1.ManagedKafka;
 import org.bf2.operator.resources.v1alpha1.ManagedKafkaAgent;
-import org.bf2.sync.client.ManagedKafkaResourceClient;
 import org.bf2.sync.informer.LocalLookup;
 
 import io.fabric8.kubernetes.client.KubernetesClient;
